sx1509: Allow using IO expander for buttons
Allow using SX1509 GPIO pins as input pins for buttons / filament
sensors / etc.

The implementation relies on an interrupt pin provided by the SX1509 to
detect change in the input pins. Once change is detected, Klipper would
read the state of the buttons via I2C and invoke callbacks as needed.

The board thus requires three GPIO pins to be connected to any Klipper
board: two pins for I2C (SDA and SCL) and one GPIO pin for the interrupt.

Signed-off-by: Ivan Dubrov <dubrov.ivan@gmail.com>